requirements.txt
tensorflow==2.12.0
streamlit
pillow
numpy
opencv-python
mediapipe
matplotlib
scikit-learn
streamlit-webrtc
